+++ Module/IntuneStarterKit/Public/Add-DeviceManagementScriptAssignment.ps1
@@ -1,0 +1,73 @@
+Function Add-DeviceManagementScriptAssignment() {
+    <#
+.SYNOPSIS
+This function is used to add a device configuration policy assignment using the Graph API REST interface
+.DESCRIPTION
+The function connects to the Graph API Interface and adds a device configuration policy assignment
+.EXAMPLE
+Add-DeviceConfigurationPolicyAssignment -ConfigurationPolicyId $ConfigurationPolicyId -TargetGroupId $TargetGroupId
+Adds a device configuration policy assignment in Intune
+.NOTES
+NAME: Add-DeviceConfigurationPolicyAssignment
+#>
+
+    [cmdletbinding()]
+
+    param
+    (
+        $ScriptId,
+        $TargetGroupId
+    )
+
+    $graphApiVersion = "Beta"
+    $Resource = "deviceManagement/deviceManagementScripts/$ScriptId/assign"
+
+    try {
+
+        if (!$ScriptId) {
+
+            write-host "No Script Policy Id specified, specify a valid Script Policy Id" -f Red
+            break
+
+        }
+
+        if (!$TargetGroupId) {
+
+            write-host "No Target Group Id specified, specify a valid Target Group Id" -f Red
+            break
+
+        }
+
+        $JSON = @"
+{
+    "deviceManagementScriptGroupAssignments":  [
+        {
+            "@odata.type":  "#microsoft.graph.deviceManagementScriptGroupAssignment",
+            "targetGroupId": "$TargetGroupId",
+            "id": "$ScriptId"
+        }
+    ]
+}
+"@
+
+        $uri = "https://graph.microsoft.com/$graphApiVersion/$Resource"
+        $Method = "POST"
+        Invoke-MgGraphRequest -Method $Method -uri $uri -Body $JSON
+
+    }
+
+    catch {
+
+        $ex = $_.Exception
+        $errorResponse = $ex.Response.GetResponseStream()
+        $reader = New-Object System.IO.StreamReader($errorResponse)
+        $reader.BaseStream.Position = 0
+        $reader.DiscardBufferedData()
+        $responseBody = $reader.ReadToEnd();
+        Write-Host "Response content:`n$responseBody" -f Red
+        Write-Error "Request to $Uri failed with HTTP Status $($ex.Response.StatusCode) $($ex.Response.StatusDescription)"
+        write-host
+        break
+
+    }
+}