+++ Module/IntuneStarterKit/Public/Add-ISK.ps1
@@ -30,7 +30,8 @@ function Add-ISK {
 
     try{
         Write-Verbose "Calling Add-ISKAPGroup for basic Groups ..."
-        $APGroup = Add-ISKGroup -GroupName $APGroupName -GroupDescription "Group containing all Autopilot registered devices" -GroupType "DynamicMembership" -GroupRule "(device.devicePhysicalIds -any _ -eq ""[OrderID]:$OrderID"")"
+        $APGroup = Add-ISKGroup -GroupName $APGroupName -GroupDescription "Group containing all Autopilot registered devices" -GroupType "DynamicMembership" -GroupRule '(device.devicePhysicalIDs -any (_ -contains "[ZTDID]"))'
+        Write-Host "Dynamic Autopilot group created: $APGroupName" -ForegroundColor Green
         #$StdGroup = Add-ISKGroup -GroupName $StdGroupName -GroupDescription "Group for standard policies and applications" -Members $APGroup
 
         Write-Verbose "Calling Add-ISKAPProfile ..."