+++ Module/IntuneStarterKit/ReleaseNotes.md
@@ -1,7 +1,8 @@
 # Release notes for IntuneStartKit module
 ## 0.1.1
 - optimisation custom config & apps
-- assignments
+- assignments for policies
+- better feedback / Verbose
 
 ## 0.1.0
 - Initial release, se README.md for documentation.